refactor: util과 컴포넌트 의존성 분리
