+++ src/components/movie/LoadMoreButton.ts
@@ -1,6 +1,10 @@
 import { MovieResponse } from "../../../types/movie";
+import { $ } from "../../utils/dom.ts";
 import Button from "../common/Button.ts";
+import hideSkeleton from "../utils/hideSkeleton.ts";
+import showSkeleton from "../utils/showSkeleton.ts";
 import MovieList from "./MovieList.ts";
+import NoSearchResults from "./NoSearchResults.ts";
 
 type Props = {
   loadFn: (currentPage: number) => Promise<MovieResponse>;
@@ -13,9 +17,18 @@ const LoadMoreButton = ({ loadFn }: Props) => {
     text: "더보기",
     className: ["load-more"],
     onClick: async () => {
-      currentPage++;
-      const movies = await loadFn(currentPage);
-      MovieList(movies);
+      showSkeleton();
+      try {
+        currentPage++;
+        const movies = await loadFn(currentPage);
+        MovieList(movies);
+      } catch (e) {
+        $("#wrapper").appendChild(
+          NoSearchResults("영화 목록을 가져오지 못했습니다.")
+        );
+      } finally {
+        hideSkeleton();
+      }
     },
   });
