+++ src/components/utils/hideSkeleton.ts
@@ -1,0 +1,7 @@
+import { $$ } from "../../utils/dom";
+
+const hideSkeleton = () => {
+  $$(".skeleton")?.forEach((s) => s.remove());
+};
+
+export default hideSkeleton;
